+++ src/graphql/resolvers.ts
@@ -52,6 +52,8 @@ import {
   MainBot,
   InputRequest,
   UserSchema,
+  HedgeComboBacktestingResult,
+  HedgeDCABacktestingResult,
 } from '../../types'
 import BotInstance from '../bot'
 import utils, { isFutures } from '../utils'
@@ -90,6 +92,8 @@ import {
   globalVarsDb,
   hedgeDCABotDb,
   userDb as _userDb,
+  hedgeComboBacktestDb,
+  hedgeDcaBacktestDb,
 } from '../db/dbInit'
 import { errorAccess } from './errorResponse'
 import {
@@ -3411,6 +3415,60 @@ const resolvers = <
         total: result.status === StatusEnum.ok ? result.data.count : 0,
       }
     },
+    getHedgeComboBacktests: async (
+      _parent: any,
+      { input }: { input?: DataGridFilterInput },
+      { token, req }: InputRequest,
+    ) => {
+      if (!req.user?.authorized) {
+        return errorAccess()
+      }
+      const user = await findUser(token)
+      if (user.status === StatusEnum.notok) {
+        return user
+      }
+      const { filter, ...rest } = mapDataGridOptionsToMongoOptions(input)
+      const result = await hedgeComboBacktestDb.readData(
+        { userId: user.data._id.toString(), ...filter },
+        undefined,
+        { ...rest },
+        true,
+        true,
+      )
+      return {
+        status: result.status,
+        data: result.status === StatusEnum.ok ? result.data.result : null,
+        reason: result.status === StatusEnum.ok ? null : result.reason,
+        total: result.status === StatusEnum.ok ? result.data.count : 0,
+      }
+    },
+    getHedgeDCABacktests: async (
+      _parent: any,
+      { input }: { input?: DataGridFilterInput },
+      { token, req }: InputRequest,
+    ) => {
+      if (!req.user?.authorized) {
+        return errorAccess()
+      }
+      const user = await findUser(token)
+      if (user.status === StatusEnum.notok) {
+        return user
+      }
+      const { filter, ...rest } = mapDataGridOptionsToMongoOptions(input)
+      const result = await hedgeDcaBacktestDb.readData(
+        { userId: user.data._id.toString(), ...filter },
+        undefined,
+        { ...rest },
+        true,
+        true,
+      )
+      return {
+        status: result.status,
+        data: result.status === StatusEnum.ok ? result.data.result : null,
+        reason: result.status === StatusEnum.ok ? null : result.reason,
+        total: result.status === StatusEnum.ok ? result.data.count : 0,
+      }
+    },
     getLeverageBracketsByUUID: async (
       _parent: any,
       { input: { uuid } }: { input: { uuid: string } },
@@ -3461,6 +3519,32 @@ const resolvers = <
       { input }: { input: { shareId: string } },
     ) => {
       const result = await comboBacktestDb.readData({ shareId: input.shareId })
+      return {
+        status: result.status,
+        data: result.status === StatusEnum.ok ? result.data.result : null,
+        reason: result.status === StatusEnum.ok ? null : result.reason,
+      }
+    },
+    getHedgeComboBacktestByShareId: async (
+      _parent: any,
+      { input }: { input: { shareId: string } },
+    ) => {
+      const result = await hedgeComboBacktestDb.readData({
+        shareId: input.shareId,
+      })
+      return {
+        status: result.status,
+        data: result.status === StatusEnum.ok ? result.data.result : null,
+        reason: result.status === StatusEnum.ok ? null : result.reason,
+      }
+    },
+    getHedgeDCABacktestByShareId: async (
+      _parent: any,
+      { input }: { input: { shareId: string } },
+    ) => {
+      const result = await hedgeDcaBacktestDb.readData({
+        shareId: input.shareId,
+      })
       return {
         status: result.status,
         data: result.status === StatusEnum.ok ? result.data.result : null,
@@ -6787,6 +6871,60 @@ const resolvers = <
         reason: result.status === StatusEnum.ok ? null : result.reason,
       }
     },
+    saveHedgeComboBacktest: async (
+      _parent: any,
+      {
+        input,
+      }: {
+        input: HedgeComboBacktestingResult
+      },
+      { token, req }: InputRequest,
+    ) => {
+      if (!req.user?.authorized) {
+        return errorAccess()
+      }
+      const user = await findUser(token)
+      if (user.status === StatusEnum.notok) {
+        return user
+      }
+      const result = await hedgeComboBacktestDb.createData({
+        ...input,
+        savePermanent: !!input.savePermanent,
+      })
+      return {
+        status: result.status,
+        data:
+          result.status === StatusEnum.ok ? result.data._id.toString() : null,
+        reason: result.status === StatusEnum.ok ? null : result.reason,
+      }
+    },
+    saveHedgeDCABacktest: async (
+      _parent: any,
+      {
+        input,
+      }: {
+        input: HedgeDCABacktestingResult
+      },
+      { token, req }: InputRequest,
+    ) => {
+      if (!req.user?.authorized) {
+        return errorAccess()
+      }
+      const user = await findUser(token)
+      if (user.status === StatusEnum.notok) {
+        return user
+      }
+      const result = await hedgeDcaBacktestDb.createData({
+        ...input,
+        savePermanent: !!input.savePermanent,
+      })
+      return {
+        status: result.status,
+        data:
+          result.status === StatusEnum.ok ? result.data._id.toString() : null,
+        reason: result.status === StatusEnum.ok ? null : result.reason,
+      }
+    },
     saveGridBacktest: async (
       _parent: any,
       {
@@ -6858,6 +6996,60 @@ const resolvers = <
         return user
       }
       const result = await comboBacktestDb.deleteManyData({
+        _id: { $in: input.ids },
+        userId: user.data._id.toString(),
+      })
+      return {
+        status: result.status,
+        data:
+          result.status === StatusEnum.ok ? 'Backtest result deleted' : null,
+        reason: result.status === StatusEnum.ok ? null : result.reason,
+      }
+    },
+    deleteHedgeComboBacktests: async (
+      _parent: any,
+      {
+        input,
+      }: {
+        input: { ids: string[] }
+      },
+      { token, req }: InputRequest,
+    ) => {
+      if (token === 'demo' || !req.user?.authorized) {
+        return errorAccess()
+      }
+      const user = await findUser(token)
+      if (user.status === StatusEnum.notok) {
+        return user
+      }
+      const result = await hedgeComboBacktestDb.deleteManyData({
+        _id: { $in: input.ids },
+        userId: user.data._id.toString(),
+      })
+      return {
+        status: result.status,
+        data:
+          result.status === StatusEnum.ok ? 'Backtest result deleted' : null,
+        reason: result.status === StatusEnum.ok ? null : result.reason,
+      }
+    },
+    deleteHedgeDCABacktests: async (
+      _parent: any,
+      {
+        input,
+      }: {
+        input: { ids: string[] }
+      },
+      { token, req }: InputRequest,
+    ) => {
+      if (token === 'demo' || !req.user?.authorized) {
+        return errorAccess()
+      }
+      const user = await findUser(token)
+      if (user.status === StatusEnum.notok) {
+        return user
+      }
+      const result = await hedgeDcaBacktestDb.deleteManyData({
         _id: { $in: input.ids },
         userId: user.data._id.toString(),
       })
@@ -6963,6 +7155,74 @@ const resolvers = <
         reason: result.status === StatusEnum.ok ? null : result.reason,
       }
     },
+    setHedgeComboBacktestPermanentStatus: async (
+      _parent: any,
+      {
+        input,
+      }: {
+        input: { id: string; savePermanent: boolean }
+      },
+      { token, req }: InputRequest,
+    ) => {
+      if (token === 'demo' || !req.user?.authorized) {
+        return errorAccess()
+      }
+      const user = await findUser(token)
+      if (user.status === StatusEnum.notok) {
+        return user
+      }
+      const { id, savePermanent } = input
+      const result = await hedgeComboBacktestDb.updateData(
+        {
+          _id: id,
+        },
+        { $set: { savePermanent: !!savePermanent } },
+      )
+      return {
+        status: result.status,
+        data:
+          result.status === StatusEnum.ok
+            ? savePermanent
+              ? 'Backtest saved permanently'
+              : `Backtest will be scheduled for deletion`
+            : null,
+        reason: result.status === StatusEnum.ok ? null : result.reason,
+      }
+    },
+    setHedgeDCABacktestPermanentStatus: async (
+      _parent: any,
+      {
+        input,
+      }: {
+        input: { id: string; savePermanent: boolean }
+      },
+      { token, req }: InputRequest,
+    ) => {
+      if (token === 'demo' || !req.user?.authorized) {
+        return errorAccess()
+      }
+      const user = await findUser(token)
+      if (user.status === StatusEnum.notok) {
+        return user
+      }
+      const { id, savePermanent } = input
+      const result = await hedgeDcaBacktestDb.updateData(
+        {
+          _id: id,
+        },
+        { $set: { savePermanent: !!savePermanent } },
+      )
+      return {
+        status: result.status,
+        data:
+          result.status === StatusEnum.ok
+            ? savePermanent
+              ? 'Backtest saved permanently'
+              : `Backtest will be scheduled for deletion`
+            : null,
+        reason: result.status === StatusEnum.ok ? null : result.reason,
+      }
+    },
     setBacktestTextFields: async (
       _parent: any,
       {
@@ -6992,10 +7252,18 @@ const resolvers = <
           ? backtestDb
           : type === BotType.combo
             ? comboBacktestDb
-            : gridBacktestDb
+            : type === BotType.hedgeCombo
+              ? hedgeComboBacktestDb
+              : type === BotType.hedgeDca
+                ? hedgeDcaBacktestDb
+                : gridBacktestDb
       const $set: { [x: string]: string | undefined } = {}
       if (typeof name !== 'undefined') {
-        $set['settings.name'] = name
+        if (type === BotType.hedgeDca || type === BotType.hedgeCombo) {
+          $set['long.settings.name'] = name
+        } else {
+          $set['settings.name'] = name
+        }
       }
       if (typeof note !== 'undefined') {
         $set.note = note
@@ -7343,6 +7611,82 @@ const resolvers = <
       }
       const shareId = v4()
       const result = await comboBacktestDb.updateData(filter, {
+        $set: { shareId },
+      })
+      return {
+        status: result.status,
+        data: result.status === StatusEnum.ok ? shareId : null,
+        reason: result.status === StatusEnum.ok ? null : result.reason,
+      }
+    },
+    shareHedgeComboBacktest: async (
+      _parent: any,
+      {
+        input,
+      }: {
+        input: { _id: string }
+      },
+      { token, req }: InputRequest,
+    ) => {
+      if (token === 'demo' || !req.user?.authorized) {
+        return errorAccess()
+      }
+      const user = await findUser(token)
+      if (user.status === StatusEnum.notok) {
+        return user
+      }
+      const filter = { _id: input._id, userId: user.data._id.toString() }
+      const get = await hedgeComboBacktestDb.readData(filter)
+      if (get.status === StatusEnum.notok) {
+        return get
+      }
+      if (get.data.result.shareId) {
+        return {
+          status: StatusEnum.ok,
+          data: get.data.result.shareId,
+          reason: null,
+        }
+      }
+      const shareId = v4()
+      const result = await hedgeComboBacktestDb.updateData(filter, {
+        $set: { shareId },
+      })
+      return {
+        status: result.status,
+        data: result.status === StatusEnum.ok ? shareId : null,
+        reason: result.status === StatusEnum.ok ? null : result.reason,
+      }
+    },
+    shareHedgeDCABacktest: async (
+      _parent: any,
+      {
+        input,
+      }: {
+        input: { _id: string }
+      },
+      { token, req }: InputRequest,
+    ) => {
+      if (token === 'demo' || !req.user?.authorized) {
+        return errorAccess()
+      }
+      const user = await findUser(token)
+      if (user.status === StatusEnum.notok) {
+        return user
+      }
+      const filter = { _id: input._id, userId: user.data._id.toString() }
+      const get = await hedgeDcaBacktestDb.readData(filter)
+      if (get.status === StatusEnum.notok) {
+        return get
+      }
+      if (get.data.result.shareId) {
+        return {
+          status: StatusEnum.ok,
+          data: get.data.result.shareId,
+          reason: null,
+        }
+      }
+      const shareId = v4()
+      const result = await hedgeDcaBacktestDb.updateData(filter, {
         $set: { shareId },
       })
       return {